% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLocations.R
\name{findLocations}
\alias{findLocations}
\title{findLocations}
\usage{
findLocations(shape, location_coordinates, location_id_colname)
}
\arguments{
\item{shape}{Shiny input (input$MAPID_draw_new_feature), representing shape
drawn on the map by the user.}

\item{location_coordinates}{A SF point object or SpatialPointsDataframe
containing coordinates and ids for all map locations.}

\item{location_id_colname}{Column name from location_coordinates containing
desired names or ids for set of locations returned.}
}
\value{
A vector of location ids.
}
\description{
Find locations inside a polygon, square, or circle drawn with leaflet.extras
drawing tools on a Shiny Leaflet map. Works with input locations as a
Spatial Points Dataframe or a Simple Features Dataframe
}
